+++ lib/rubocop/processed_source.rb
@@ -31,10 +31,6 @@ module RuboCop
       @comment_config ||= CommentConfig.new(self)
     end
 
-    def disabled_line_ranges
-      comment_config.cop_disabled_line_ranges
-    end
-
     def lines
       @lines ||= raw_source.lines.map(&:chomp)
     end